Renamed Protocols to QueryClients
